Move deck builder to its own Vuex module
The current deck builder reuses a bunch of stuff from the existing
replay logic: a bunch of vuex state as well as the DraftTable component.

In order to url-ify the replay logic, we can't really share this stuff
anymore -- the replay logic has become too intertwined with its own url
management.

So instead we do two things:
[1] Spin all of the deck builder state off into its own Vuex module
[2] Replace the usage of DraftTable with DeckBuilderPlayerSelector,
    a new component (that's also dead simple).