Clean up routing a bit
Rename Home -> Replay
Rename DraftBuilderScreen -> DeckBuilder

Change Replay so it only has one entry in the routing. Modify the
URL parsing so we just accept arbitrary params after /:draftId and do
the parsing manually.

If we can get it so that each root component only has one entry in the
routes then we can, in theory, do auto-code splitting. As soon as I figure
out how that works.