[replay] Open packs as soon as previous pack empties
Previously, we waited for an event from the next round before pushing the
"open the next pack" event to stream. However, this caused weird situations
where a player may have packs queued from the next round without having
yet opened their pack for that round.

This approach changes things so that a player opens the next pack as soon
as they pick a card from a pack with only one card in it.

This doesn't 100% fix the problem -- it's still possible for player to
have a queue of [(round 2, pick 15), (round 3, pick 2)] but this ensures
that the "head" of the queue always represents the pack that the user
will pick from next.