Fix replay
This involves a lot of refactors that I'm too tired to explain, but in
summary:

- Generalize "pack locations" to places that packs can be. Such as:
  [a] A player's opened packs
  [b] A player's unopened packs
  [c] Packs not in the cube
  [d] The pack graveyard (where empty packs go)
  Every pack location has its own ID and can be referred to diretly by
  mutations.

- Be extra careful about where packs end up going when they're passed.
  It's not safe to always push them to the end of the next player's
  queue because that player might have packs from the next round already
  queued.

There's still a decent amount of information erasure taking place because
we're not storing the index that packs are coming from or going to. As a
result, we have make the mutation system do a lot more work than it should.

Long term, we should generate two separate event streams for timeline mode
and synchronized mode that have different resolved indices.