Ensure we update state maps every time
The draft state is essentially two pieces:
[1] The actual state itself
[2] Maps into particular categories of that state. For example, we have
    one map of all the packs in the draft.

These maps need to be generated when the state is first created and
regenerated when the state is cloned. Previously we were doing those steps
in different places, which was causing bugs. Now that's all centralized
into fillDraftStateMaps().

Also makes it so that DraftCard.pickedIn no longer references the
actual event in which it was picked but a data blob that describes the
event. Object references like that don't work well given how we clone the
draft state.